prevent excessive backtracking in regex engine
The following font lock regex pattern leads to excessive backtracking
due to an arbitrary number of characters in the parenthesis and an
arbitrary number of parenthesis itself:

"^\\(?:.+?-\\)*?[0-9]+-.*\n"

This leads to a lot of backtracking. The time required for matching
grows exponentially with the input length (the line length in rg's
case).

Fixes #14. It might also be the cause of #12.