Only append '.' to command line on windows
rg changed it output in version 0.10.0 to include current dir in file paths
in case it is explicitly entered on command line. This looks ugly so
let's just use that hack on windows where it, for some reason, is
needed.

This also triggered some fixing of test cases and helpers to avoid
problems.