Make column numbers invisible when customized to not show them
Visible column number information is seldom helpful in the results
buffer but we want to be able to jump to the right column when
clicking on a search result. Compilation mode needs the column numbers
in the buffer to be able to jump to the correct place so always
include them but make them invisible if rg-show-columns is nil.
This allows for a cleaner look while still take advantage of precision
jumps.