Unify toggle flags and other command line flags in rg-search struct
The separation of toggle flags causes harm in relation to the rg-menu
when flags set there that are also toggle flags. These
will not be possible to toggle and the UX shows wrong indications for
case and ignore state.

The separation doesn't provide much benefit except for catching
potential logical errors in the program itself or related toggle
definitions.

This patch unifies all the flags into one field to allow correct
interaction with the rg-menu.