Set the compilation-error-regexp-alist in a hook (#126)
Some modes, for instance verilog-mode, adds to the
compilation-error-regexp-alist and friends which could interfere with
rg-mode regexps. Since rg-mode isn't a generic compilation mode we
want to avoid that.
This solution defer setting these regexps to the last run
compilation-mode-hook to be able to override whatever happens in other
hooks.