Fix "File: " skipping when building the elements of the imenu
We shouldn't be using `skip-chars-forward' here as this function skips
over any _character set_ passed as argument, not just a string.

Without this patch, for instance entries like:

  File: lol/foo.bar

Will be added to the candidates list as "ol/foo.bar" as ?l is in the
character set "File: ".

Instead, just skips 6 chars forward.