Adapt header line mouse clicks to emacs 30.2 (#196)
A bug was fixed in emacs 30.2 so that we no longer need to
use the wrong mouse button in the header line key map.
Since this is a breaking change we need to keep the old
mapping for older emacsen.

So for emacs < 30.2 we use mouse-2 and for emacs >= 30.2
we use mouse-1.